Separated tests cases of login and home pages
